+++ nengo/networks/tests/test_ensemblearray.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 import nengo
-from nengo.dists import Choice
+from nengo.dists import Choice, UniformHypersphere
 from nengo.exceptions import ValidationError
 from nengo.utils.compat import range
 
@@ -238,3 +238,41 @@ def test_neuronoutput(Simulator, seed):
     with Simulator(net) as sim:
         sim.run(0.01)
     assert np.all(sim.data[p] < 1e-2)
+
+
+def test_ndarrays(Simulator, rng):
+    encoders = UniformHypersphere(surface=True).sample(10, 1, rng=rng)
+    max_rates = rng.uniform(200, 400, size=10)
+    intercepts = rng.uniform(-1, 1, size=10)
+    eval_points = rng.uniform(-1, 1, size=(100, 1))
+    with nengo.Network() as net:
+        ea = nengo.networks.EnsembleArray(
+            10, 2,
+            encoders=encoders,
+            max_rates=max_rates,
+            intercepts=intercepts,
+            eval_points=eval_points,
+            n_eval_points=eval_points.shape[0])
+
+    with Simulator(net) as sim:
+        pass
+    built = sim.data[ea.ea_ensembles[0]]
+    assert np.allclose(built.encoders, encoders)
+    assert np.allclose(built.max_rates, max_rates)
+    assert np.allclose(built.intercepts, intercepts)
+    assert np.allclose(built.eval_points, eval_points)
+    assert built.eval_points.shape == eval_points.shape
+
+    with nengo.Network() as net:
+        # Incorrect shapes don't fail until build
+        ea = nengo.networks.EnsembleArray(
+            10, 2,
+            encoders=encoders,
+            max_rates=max_rates,
+            intercepts=intercepts,
+            eval_points=eval_points[:10],
+            n_eval_points=eval_points.shape[0])
+
+    with pytest.raises(ValidationError):
+        with Simulator(net) as sim:
+            pass